+++ dart/test/protocol/sentry_feature_flags_tests.dart
@@ -7,16 +7,16 @@ import '../mocks.dart';
 void main() {
   final featureFlags = SentryFeatureFlags(
     values: [
-      SentryFeatureFlag(name: 'feature_flag_1', value: 'value_1'),
-      SentryFeatureFlag(name: 'feature_flag_2', value: 'value_2'),
+      SentryFeatureFlag(name: 'feature_flag_1', value: true),
+      SentryFeatureFlag(name: 'feature_flag_2', value: false),
     ],
     unknown: testUnknown,
   );
   final featureFlagsJson = <String, dynamic>{
     ...testUnknown,
     'values': [
-      {'name': 'feature_flag_1', 'value': 'value_1'},
-      {'name': 'feature_flag_2', 'value': 'value_2'},
+      {'name': 'feature_flag_1', 'value': true},
+      {'name': 'feature_flag_2', 'value': false},
     ],
   };